move builder code to extra file
